+++ tests/unit/db-versionManager.test.ts
@@ -64,8 +64,7 @@ afterEach(() => {
 });
 
 after(() => {
-  if (fs.existsSync(fileTmpDir))
-    fs.rmSync(fileTmpDir, { recursive: true, force: true, maxRetries: 5, retryDelay: 100 });
+  if (fs.existsSync(fileTmpDir)) fs.rmSync(fileTmpDir, { recursive: true, force: true, maxRetries: 5, retryDelay: 100 });
 });
 
 async function resetModuleStorage() {
